Rewrite the views system to use xrpc
