+++ packages/server/src/api/index.ts
@@ -1,16 +1,10 @@
 import { createServer } from '../xrpc'
-import names from './names'
-import session from './session'
-import account from './account'
-import repo from './repo'
-import sync from './sync'
+import todoAdx from './todo/adx'
+import todoSocial from './todo/social'
 
 export default function () {
   const server = createServer()
-  names(server)
-  session(server)
-  account(server)
-  repo(server)
-  sync(server)
+  todoAdx(server)
+  todoSocial(server)
   return server
 }
