+++ packages/server/src/db/types.ts
@@ -1,5 +1,4 @@
 import { AdxUri } from '@adxp/common'
-import { DataSource } from 'typeorm'
 import { AdxValidationResult } from '@adxp/schemas'
 
 export type DbRecordPlugin<T, S> = {
@@ -11,13 +10,3 @@ export type DbRecordPlugin<T, S> = {
   delete: (uri: AdxUri) => Promise<void>
   translateDbObj: (dbObj: S) => T
 }
-
-export type DbViewPlugin = {
-  id: string
-  fn: (db: DataSource) => ViewFn
-}
-
-export type ViewFn = (
-  params: Record<string, unknown>,
-  requesterDid: string,
-) => unknown
