--- packages/server/src/db/views/index.ts
@@ -1,19 +1,0 @@
-import likedBy from './likedBy'
-import repostedBy from './repostedBy'
-import userFollows from './userFollows'
-import userFollowers from './userFollowers'
-import profile from './profile'
-import feed from './feed'
-import postThread from './postThread'
-
-const views = [
-  likedBy,
-  repostedBy,
-  userFollows,
-  userFollowers,
-  profile,
-  feed,
-  postThread,
-]
-
-export default views
